LGTM

+++ src/tools/Communicator.cpp
@@ -187,7 +187,7 @@ void Communicator::Allgatherv(ConstData in,Data out,const int*recvcounts,const i
     plumed_assert(rc);
     plumed_assert(rc[0]==in.size);
     plumed_assert(di);
-    if(s) std::memcpy(static_cast<char*>(r)+displs[0]*in.nbytes,s,in.size*in.nbytes);
+    if(s) std::memcpy(static_cast<char*>(r)+displs[0]*in.nbytes,s,size_t(in.size)*in.nbytes);
   }
 #else
   plumed_assert(in.nbytes==out.nbytes);
@@ -195,7 +195,7 @@ void Communicator::Allgatherv(ConstData in,Data out,const int*recvcounts,const i
   plumed_assert(rc);
   plumed_assert(rc[0]==in.size);
   plumed_assert(di);
-  if(s) std::memcpy(static_cast<char*>(r)+displs[0]*in.nbytes,s,in.size*in.nbytes);
+  if(s) std::memcpy(static_cast<char*>(r)+displs[0]*in.nbytes,s,size_t(in.size)*in.nbytes);
 #endif
 }
 
@@ -209,12 +209,12 @@ void Communicator::Allgather(ConstData in,Data out) {
   } else {
     plumed_assert(in.nbytes==out.nbytes);
     plumed_assert(in.size==out.size);
-    if(s) std::memcpy(r,s,in.size*in.nbytes);
+    if(s) std::memcpy(r,s,size_t(in.size)*in.nbytes);
   }
 #else
   plumed_assert(in.nbytes==out.nbytes);
   plumed_assert(in.size==out.size);
-  if(s) std::memcpy(r,s,in.size*in.nbytes);
+  if(s) std::memcpy(r,s,size_t(in.size)*in.nbytes);
 #endif
 }
